Add logo to the repo
This is based on details on ADI's branding guidelines at:
https://www.analog.com/-/media/analog/en/about-adi/newsroom/multi-media-gallery/logos/awp-adi-logo-guidelines-2016.pdf

The font is AzbukaW01-Medium and AzbukaW01-Light from:
https://www.onlinewebfonts.com/download/5707abbe50e51d3cdf9d0b7abb3bdc04
https://www.onlinewebfonts.com/download/307ee86f70f31249c3480cd93ff30c5a
Authors: Richard Dawson and Dave Farey
Font is licensed by CC BY 3.
https://creativecommons.org/licenses/by/3.0/us/

This was approved by the Python Software Foundation, per
https://www.python.org/psf/trademarks/
as it incorporates the Python logo as well.

Logo was created by me in inkscape (svg file), and pngs were exported
from that.

Signed-off-by: Robin Getz <robin.getz@analog.com>