Ad7124 (#58)
* rx_tx: disable all channels before re-enabling
* ad7124: use bigendian signed 32bit int as data type
* rx_tx: added left/right shifting to the data and masking
* ad7124: added mask and changed data type to unsigned int 32 big endian
* ad7124: sample rate is channel attribute and affects all channels
* ad7124: added automatic part detection.
When creating a new ad7124 object pyadi-iio will now look for both -8 and -4 device types.
It will select the device_index-th connected device from the 7124 family as working device. device_index defaults to 0.
* ad7124: sort device channels after the first voltage channel's index
The available devices were lexicographical ordered. This was confusing when enabling the channels using their indexes.
Now we are sorting the available channels after the index of the first voltage channel appearing in their names.
* ad7124: changed non_complex data to millivolts.

Signed-off-by: Cristi Iacob <cristian.iacob@analog.com>
Signed-off-by: Adrian Suciu <adrian.suciu@analog.com>