examples/QuadMxFE_example: Add QuadMxFE Example
This example requires the external calibration board attached, in order
to loop the Combined Tx Channels Back Into Combined Rx Path.
One TX is configured to generate a single tone of fsamp/50.
Then we enable the first Rx channel of each MxFE and, use the first
as reference and compute the phase offset between the remaining MxFEs.
Between runs the all MxFEs are resynced and the JESD links are brought
up and down. The expected result is a deterministic phase offset
between iterations and runs.

Signed-off-by: Michael Hennerich <michael.hennerich@analog.com>