test: adrv9002: update tests to new api
The main difference with the new API is that it's made clear that the RX
gain tables type cannot really change at runtime. Hence, the driver was
fixed accordingly which means a different ABI for the digital gain
attributes. This change reflects that...

Also to note that for now the driver does not support compensated gain
tables because there's still no way to get that information from the TES
generated profiles (hopefully we'll get that in the next SDK version).
Hence all the tests assume that correction tables are being used which
means that all the stuff for the compensated tables was removed.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>