tests: dma_tests: fix data capture for low speed devices
By default, the dma_rx function is setting the number of samples to
capture to 2^15. This is a very high number for low speed devices and
for those devices, libiio will raise a timeout.

This change makes the number of samples to capture a parameter of the
function and keeps the default value to 2^15.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>