add support for adg2128
ADG2128 is an 12x8 cross point switch. As it stands, IIO does not
have any support for cross point switches. So in order to have a
way to drive this chip using libiio, this commit adds a driver
implementation for adg2128 written in python that makes use of the
direct register access feature of IIO to perform I2C read/write
operations.

Ideally, IIO should be extended to support such devices (cross
point switches) but until this happens, this seems to be an
acceptable compromise to be able to communicate with adg2128
device in the same IIO connection with other ADI devices.

examples/adg2128.py shows how the driver exposes a 12x8 2D
array-like access to the switches to the user:
device[x][y] = True/False # closed/open

The driver supports communicating with several adg2128 devices
on the same bus, with common Y connections and all this is
presented to the user in a larger 2D array:
- 24x8 for 2 devices
- 36x8 for 3 devices
...

Signed-off-by: Darius Berghe <darius.berghe@analog.com>