CN0511 calibration support
The calibration of cn0511 output was added in adi/cn0511.py.
An example of how to set the callibrated output was added in
examples/cn0511.py.
If libad9166 cannot be imported, the calibration will not work, some
warnings will be printed and the examples/cn0511_example.py will run
accordingly.

Signed-off-by: Beleca <Valentin.Beleca@analog.com>