rx_tx: use __rx_buffered_data() to acquire samples
The __rx_buffered_data method uses iio channel.read() method which handles
endianness and shift. After this the data is converted using the
np.frombuffer method. Data is converted on the spot so multi-type buffers
are handled immediatly. One side-effect of using channel.read is that
data is channel interleaved as opposed to sample interleaved, so we
need to take care of that.

Eventually this method is used for both complex and non-complex acquisition

Signed-off-by: Adrian Suciu <adrian.suciu@analog.com>