Add delay and disable shVDD detection
The AD74413R asserts the fault bits after a 2 ms debounce period. Add
a delay between setting an output value and reading the alert status
register.

If shVDD detection is enabled, configuring a channel as high impedance
will result in outputing 24V, which will cause the loopback tests to
fail. So, disable shVDD detection after the test is done.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>