adi/ad2s1210: add AD2S1210 Resolver-to-Digital Converter
This adds support for the AD2S1210 Resolver-to-Digital Converter.

The IIO ABI is based on work that is currently being mainlined and
is not compatible with the older driver in staging.

Not all features are supported yet since libiio does not support iio
events. This means that the AD2S1210's fault detection and threshold
configuration is not supported.

The tests are passing when run on real hardware.

Signed-off-by: David Lechner <dlechner@baylibre.com>