ad7606: implement per-channel scale-available support
The AD7606C-16 and AD7606C-18 support per-channel 'scale_availble' in SW
mode.
That is because each channel can be configured individually, to be
bipolar-differential, bipolar-single-ended or unipolar-single-ended.

Depending on the above configuration, the available scales change.
So, we need to tweak this per-channel.

Since there isn't a way to detect SW mode, we use some 'try' blocks in
Python.

Signed-off-by: Alexandru Ardelean <aardelean@baylibre.com>