jupiter_sync.py: Updates according to the newer Linux driver version
Remove the need of moving all channels to calibrated state, before the
request for the device to enter in MCS state. The Linux driver will do
this automatically, furthermore, it will restore the previous state
of the channel before MCS.

Simplification for the internal/external sync logic.
The Internal sync will only be used when having to synchronize the
channels of one system. Because of this, the updated version of
the Linux driver, will automatically trigger the MCS procedure,
including the internal sync, after a profile is loaded.
The external sync will be required when dealing with multiple systems.
The information regarding the need to sync one or more systems is
defined in the devicetree.

The Linux driver exposes the "multi_chip_sync" attribute, which is a bit
more special.
To know that the MCS was done successfully, one has to wait for the
multi_chip_sync attribute setting process to end successfully.
The attribute setting of multi_chip_sync only ends after MCS procedure is
completed or timeout occurs.
For the MCS sync to end, the adrv9002 needs to receive the 6 MCS pulse train.
In other words, we need to configure(put in MCS state) each adrv9002
in a separate thread, send the 6 MCS pulse train and wait for all started
threads that have put the adrv9002 devices in MCS state to end in success.

Signed-off-by: AndreiGrozav <andrei.grozav@analog.com>