context_manager.py: Fix error handling for usb
In the case of usb connections, no two iio.Context objects can exist at
the same time. When the `dev_interface` functions are called inside the
test functions, the function may fail before the `del sdr` operation can
be performed and the iio.Context object is not destroyed. The next test
that calls the `dev_interface` function will therefore also fail (since
it cannot create another iio.Context object).
By transforming the `context_manager` class into a Pythonic context
manager, we make use of the `with` statement. The `with` block will act
as a wrapper over the usage of iio.Context objects and it will ensure
that when errors are thrown, they are logged and memory is properly
freed.
`context_manager` objects can still be created without the `with`
statement (introduced change doesn't affect pre-existing code).

Signed-off-by: bia1708 <bianca-maria.popu@analog.com>